XIVY-13091 Cleanup Browser Mock-Tests
